+++ plugin/src/main/java/com/stratio/cassandra/lucene/schema/mapping/SingleColumnMapper.java
@@ -161,7 +161,7 @@ public abstract class SingleColumnMapper<T extends Comparable<T>> extends Mapper
      *
      * @param <T> the base type
      */
-    public static abstract class SingleFieldMapper<T extends Comparable<T>> extends SingleColumnMapper<T> {
+    public abstract static class SingleFieldMapper<T extends Comparable<T>> extends SingleColumnMapper<T> {
 
         /**
          * Builds a new {@link SingleFieldMapper} supporting the specified types for indexing and clustering.